requirements.txt
requests
flask
Pillow